Move StepNavigation, update imports and translations
