--- packages/lib-classifier/src/components/Classifier/components/SlideTutorial/components/index.js
@@ -1,1 +1,0 @@
-export { default as StepNavigation } from './StepNavigation'
