+++ packages/lib-classifier/src/components/Classifier/components/SubjectViewer/components/ImageAndTextViewer/ImageAndTextViewerContainer.js
@@ -3,10 +3,10 @@ import PropTypes from 'prop-types'
 import React from 'react'
 import asyncStates from '@zooniverse/async-states'
 
+import StepNavigation from '@shared/StepNavigation'
 import locationValidator from '../../helpers/locationValidator'
 import SingleImageViewer from '../SingleImageViewer'
 import SingleTextViewer from '../SingleTextViewer'
-import StepNavigation from '@components/Classifier/components/SlideTutorial/components/StepNavigation'
 
 const defaultDimensions = [{
   clientHeight: 400
