+++ packages/lib-classifier/src/components/Classifier/components/shared/StepNavigation/StepNavigation.js
@@ -52,12 +52,11 @@ function StepNavigation({
     const nextStep = stepIndex + 1
     const prevStep = stepIndex - 1
     const options = steps.map((step, index) => {
-      const label = t('SlideTutorial.StepNavigation.go', { index: index + 1 })
+      const label = t('StepNavigation.go', { index: index + 1 })
       // We can't just use index for the value
       // because Grommet is using indexes internally as keys and this will error with a duplicate key
       const value = `step-${index}`
       return {
-        id: value,
         label,
         value
       }
@@ -65,7 +64,7 @@ function StepNavigation({
     return (
       <Box align='center' className={className} direction='row' justify='center' margin={{ top: 'medium' }}>
         <StyledButton
-          a11yTitle={t('SlideTutorial.StepNavigation.previous')}
+          a11yTitle={t('StepNavigation.previous')}
           data-index={prevStep}
           disabled={stepIndex === 0}
           icon={<FormPrevious />}
@@ -75,13 +74,12 @@ function StepNavigation({
         <StyledRadioButtonGroup
           direction='row'
           gap='none'
-          name='step-selectors'
           onChange={onRadioChange}
           options={options}
           value={`step-${stepIndex}`}
         />
         <StyledButton
-          a11yTitle={t('SlideTutorial.StepNavigation.next')}
+          a11yTitle={t('StepNavigation.next')}
           data-index={nextStep}
           disabled={stepIndex === steps.length - 1}
           icon={<FormNext />}
