+++ packages/lib-classifier/src/translations/en/components.json
@@ -148,12 +148,12 @@
     "alt": "Tutorial step {{activeStep}}",
     "error": "Tutorial step could not be loaded.",
     "getStarted": "Get Started",
-    "heading": "{{projectDisplayName}} Tutorial",
-    "StepNavigation": {
-      "go": "Go to step {{index}}",
-      "next": "Go to next step",
-      "previous": "Go to previous step"
-    }
+    "heading": "{{projectDisplayName}} Tutorial"
+  },
+  "StepNavigation": {
+    "go": "Go to step {{index}}",
+    "next": "Go to next step",
+    "previous": "Go to previous step"
   },
   "SubjectViewer": {
     "BarChartViewer": {
